+++ functions.php
@@ -26,8 +26,21 @@ add_action( 'after_setup_theme', 'nx_theme_setup' );
 // Includes
 require_once (NX_CORE . 'theme-scripts.php');												// Styles & Scripts
 require_once (NX_CORE . 'actions.php');														// Actions & Filters
-require_once (NX_CORE . 'admin/main.php');													// Admin settings
+require_once (NX_CORE . 'wp-bootstrap-navwalker.php');										// WP Bootstrap Navwalker
 require_once (NX_CORE . 'general-functions.php');											// General functions
 require_once (NX_CORE . 'shortcode-gallery.php');											// Gallery shortcode
-require_once (NX_CORE . 'shortcode-embed.php');												// Embed shortcode
 require_once (NX_ROOT . 'widgets/main.php');												// SiteOrigin Page Builder - Custom widgets
+
+// Register the admin page
+if(function_exists('acf_add_options_page')){
+	
+	// Theme options page
+	acf_add_options_page(array(
+		'page_title'		 => 'Website settings',
+		'menu_title'		 => 'Website settings',
+		'menu_slug'			 => 'website-settings',
+		'capability'		 => 'edit_posts',
+		'redirect'			 => false
+	));
+	
+}