+++ widgets/main.php
@@ -1,43 +1,55 @@
 <?php if ( ! defined( 'ABSPATH' ) ) { exit; }
 
-/* 
- * All folders are created for Siteorigin Page Builder. 
- * If you want to create any other widgets (not for Siteorigin Page Builder), please create another folder
- * so we'll not have problems with the page builder.
- */
-
 // Page Builder widgets path
-function nx_widgets_collection($folders){
+add_filter('siteorigin_widgets_widget_folders', function($folders){
 	
 	$folders[] = get_template_directory() . '/widgets/';
+	
 	return $folders;
 	
-}
-add_filter('siteorigin_widgets_widget_folders', 'nx_widgets_collection');
+});
 
 // Create a group of widgets 
-function nx_add_widget_tabs($tabs) {
+add_filter('siteorigin_panels_widget_dialog_tabs', function($tabs) {
+	
     $tabs[] = array(
         'title' => 'Custom Widgets',
         'filter' => array(
-            'groups' => array('nx_widgets')
+			'groups' => array('nx_widgets')
         )
     );
-    return $tabs;
-}
-add_filter('siteorigin_panels_widget_dialog_tabs', 'nx_add_widget_tabs', 20);
-
-// Make sure to activate the widgets
-function nx_default_widgets( $widgets ) {
 	
-  $widgets['accordion']			 = true;
-  $widgets['lightbox-img']		 = true;
-  $widgets['slider']			 = true;
-  $widgets['modal']				 = true;
-  
-  // SiteOrigin default widgets
-  $widgets['button']			 = false;
-  
-  return $widgets;
-}
-add_filter('siteorigin_widgets_active_widgets', 'nx_default_widgets');
+    return $tabs;
+	
+}, 20);
+
+// Automatically activate the new widgets
+add_filter('siteorigin_widgets_active_widgets', function($widgets) {
+	
+	// Thew new widgets
+	$tmp_path				 = NX_ROOT . 'widgets/';
+	$init_dirs				 = glob($tmp_path . '*' , GLOB_ONLYDIR);
+
+	if(!empty($init_dirs)){
+
+		foreach($init_dirs as $tmp_dir){
+			
+			$widget_dir		 = str_replace($tmp_path, '', $tmp_dir);
+			
+			// Activate the widget
+			if(!empty($widget_dir)){
+				
+				$widgets[$widget_dir] = true;
+				
+			}
+			
+		}
+
+	}
+
+	// SiteOrigin default widgets
+	$widgets['button']			 = false;
+
+	return $widgets;
+	
+});